Use outputFilename with built-in style templates
When a build-in style template is used, the `outputFilename` is ignored and the filename is generated based on the font name. It should only generate this if a specific `outputFilename` is not set. This allows for generating a SCSS file beginning with an underscore (partial SASS file) using the build-in SCSS template.